+++ packages/calendar-input/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Change Log
 
+## 7.1.5
+
+### Patch Changes
+
+-   [#277](https://github.com/core-ds/core-components/pull/277): Внешняя ошибка больше не влияет на вызов onChange. Thanks [@reme3d2y](https://github.com/reme3d2y)
+
+## 7.1.4
+
+### Patch Changes
+
+-   Updated dependencies [[#175](https://github.com/core-ds/core-components/pull/175)]
+    -   @alfalab/core-components-calendar@6.1.4
+
 ## 7.1.3
 
 ### Patch Changes