+++ packages/intl-phone-input/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Change Log
 
+## 7.0.13
+
+### Patch Changes
+
+-   [#279](https://github.com/core-ds/core-components/pull/279): chore: обновились @alfalab-data и @alfalab/utils версии в зависимостях. Thanks [@EGNKupava](https://github.com/EGNKupava)
+
 ## 7.0.12
 
 ### Patch Changes